fork now breaks away with test/app-access-token.js
